+++ src/js_sync.hpp
@@ -233,6 +233,8 @@ public:
     static void add_connection_notification(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &);
     static void remove_connection_notification(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &);
     static void is_connected(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &);
+    static void start(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &);
+    static void stop(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &);
 
 
     static void override_server(ContextType ctx, ObjectType this_object, Arguments args, ReturnValue&);
@@ -254,6 +256,8 @@ public:
         {"addConnectionNotification", wrap<add_connection_notification>},
         {"removeConnectionNotification", wrap<remove_connection_notification>},
         {"isConnected", wrap<is_connected>},
+        {"start", wrap<start>},
+        {"stop", wrap<stop>},
     };
 
 private:
@@ -644,6 +648,24 @@ void SessionClass<T>::is_connected(ContextType ctx, FunctionType, ObjectType thi
             return_value.set(true);
         }
     }
+}
+
+template<typename T>
+void SessionClass<T>::start(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
+    validate_argument_count(argc, 0);
+    return_value.set(false);
+    if (auto session = get_internal<T, SessionClass<T>>(this_object)->lock()) {
+        session->revive_if_needed();
+    }    
+}
+
+template<typename T>
+void SessionClass<T>::stop(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
+    validate_argument_count(argc, 0);
+    return_value.set(false);
+    if (auto session = get_internal<T, SessionClass<T>>(this_object)->lock()) {
+        session->log_out();
+    }    
 }
 
 template<typename T>