+++ tests/js/session-tests.js
@@ -84,6 +84,16 @@ function runOutOfProcess() {
     });
 }
 
+function waitForSessionConnected(session) {
+    return new Promise(res => {
+        session.addConnectionNotification((newState, oldState) => {
+            if (newState === Realm.Sync.ConnectionState.Connected) {
+                res();
+            }
+        })
+    });
+}
+
 module.exports = {
     testLocalRealmHasNoSession() {
         let realm = new Realm();
@@ -1041,9 +1051,119 @@ module.exports = {
                             resolve('Done');
                         }
                     });
+                    setTimeout(() => { reject() }, 10000);
                 }).catch(error => reject(error));
             });
         });
+    },
+
+    testStartStop() {
+        if(!isNodeProccess) {
+            return;
+        }
+
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password')
+        .then((user) => {
+            const config = {
+                sync: {
+                    user: user,
+                    url: `realm://localhost:9080/~/${uuid()}`,
+                    fullSynchronization: true,
+                }
+            };
+
+            return Realm.open(config);
+        }).then((realm) => {
+            return new Promise((resolve, reject) => {
+                const session = realm.syncSession;
+
+                const checks = {
+                    started: false,
+                    stopped: false,
+                    restarted: false
+                }
+
+                session.addConnectionNotification((newState, oldState) => {
+                    if (newState === Realm.Sync.ConnectionState.Connected && checks.started === false) { checks.started = true; session.stop(); }
+                    if (newState === Realm.Sync.ConnectionState.Connected && checks.started === true) { checks.restarted = true; resolve(); }
+                    if (newState === Realm.Sync.ConnectionState.Disconnected) { checks.stopped = true; session.start();}
+                });
+
+                setTimeout(() => { reject("Timeout") }, 10000);
+            })
+        })
+    },
+
+    testMultipleStarts() {
+        if(!isNodeProccess) {
+            return;
+        }
+
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password')
+        .then((user) => {
+            const config = {
+                sync: {
+                    user: user,
+                    url: `realm://localhost:9080/~/${uuid()}`,
+                    fullSynchronization: true,
+                }
+            };
+
+            return Realm.open(config)
+        }).then(realm => {
+            return new Promise((resolve, reject) => {
+                const session = realm.syncSession;
+
+                waitForSessionConnected(session).then(() => {
+                    session.start();
+                    session.start();
+                    session.start();
+                    setTimeout(() => {
+                        if (session.isConnected()) {
+                            resolve();
+                        } else {
+                            reject();
+                        }
+                    }, 1000); 
+                })
+            })
+        })
+    },
+
+    testMultipleStopped() {
+        if(!isNodeProccess) {
+            return;
+        }
+
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password')
+        .then((user) => {
+            const config = {
+                sync: {
+                    user: user,
+                    url: `realm://localhost:9080/~/${uuid()}`,
+                    fullSynchronization: true,
+                }
+            };
+
+            return Realm.open(config)
+        }).then(realm => {
+            return new Promise((resolve, reject) => {
+                const session = realm.syncSession;
+
+                waitForSessionConnected(session).then(() => {
+                    session.stop();
+                    session.stop();
+                    session.stop();
+                    setTimeout(() => {
+                        if (session.isConnected()) {
+                            reject();
+                        } else {
+                            resolve();
+                        }
+                    }, 1000);                       
+                })
+            })
+        })
     },
 
     testOfflinePermissionSchemas() {